refactor: sorted string
