+++ src/common/utils/index.ts
@@ -2,5 +2,6 @@ export { validateFile } from './validate-file.ts'
 export { validateImage } from './validate-image.ts'
 export { formatTextBr } from './format-text-br.tsx'
 export * from './zod-schemas.ts'
-export { errorNotification } from './errorNotification.ts'
+export { errorNotification } from './error-notification.ts'
 export { requestHandler } from './request-handler.ts'
+export { getSortedString } from './get-sorted-string.ts'
